fix : 일부 수정
